+++ lib/app.js
@@ -2,8 +2,8 @@ require('dotenv').config()
 const express = require('express')
 const bodyParser = require('body-parser')
 const mongoose = require('mongoose')
-const gameRouter = require('./application/game/gameRouter')
-const userRouter = require('./application/user/userRouter')
+const gameRouter = require('./application/game/')
+const userRouter = require('./application/user/')
 const mongoConf = process.env.MONGO_CONFIG_URL
 const path = require('path')
 const app = express()
