+++ lib/application/game/index.js
@@ -1,6 +1,6 @@
 const express = require('express')
 const router = express.Router()
-const GameController = require('./gameController')
+const GameController = require('./game-controller')
 
 router.get('/', GameController.getGames)
