+++ lib/application/user/index.js
@@ -1,6 +1,6 @@
 const express = require('express')
 const router = express.Router()
-const UserController = require('./userController')
+const UserController = require('./user-controller')
 
 router.post('/search', UserController.searchUsers)
